requirements.txt
yagmail
requests
baidu-aip
beautifulsoup4
html5lib